src/shape_generator/ml_generator.py
"""
Text2PointCloud - Machine Learning Point Cloud Generation

This module handles ML-based point cloud generation from text descriptions.
"""

import numpy as np
import torch
import torch.nn as nn
import torch.nn.functional as F
from typing import List, Tuple, Dict
import json
import os

class PointCloudGenerator(nn.Module):
    """
    Neural network for generating point clouds from text embeddings.
    """
    
    def __init__(self, text_dim=512, point_dim=3, num_points=1024):
        super().__init__()
        self.text_dim = text_dim
        self.point_dim = point_dim
        self.num_points = num_points
        
        # Text processing layers
        self.text_encoder = nn.Sequential(
            nn.Linear(text_dim, 256),
            nn.ReLU(),
            nn.Linear(256, 128),
            nn.ReLU(),
            nn.Linear(128, 64)
        )
        
        # Point cloud generation layers
        self.point_decoder = nn.Sequential(
            nn.Linear(64, 128),
            nn.ReLU(),
            nn.Linear(128, 256),
            nn.ReLU(),
            nn.Linear(256, 512),
            nn.ReLU(),
            nn.Linear(512, num_points * point_dim)
        )
        
    def forward(self, text_embedding):
        """
        Generate point cloud from text embedding.
        
        Args:
            text_embedding: Text embedding tensor
            
        Returns:
            Point cloud tensor of shape (batch_size, num_points, point_dim)
        """
        # Encode text
        encoded_text = self.text_encoder(text_embedding)
        
        # Generate point cloud
        point_cloud = self.point_decoder(encoded_text)
        
        # Reshape to point cloud format
        point_cloud = point_cloud.view(-1, self.num_points, self.point_dim)
        
        return point_cloud

class TextToPointCloudML:
    """
    Main class for ML-based text-to-point-cloud generation.
    """
    
    def __init__(self):
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        self.model = None
        self.text_encoder = None
        self.load_or_initialize_model()
    
    def load_or_initialize_model(self):
        """Load pre-trained model or initialize new one."""
        # For now, initialize a new model
        # In practice, you'd load a pre-trained model
        self.model = PointCloudGenerator().to(self.device)
        print(f"Initialized model on {self.device}")
    
    def generate_point_cloud(self, text_input: str, num_points: int = 1024) -> List[Tuple[float, float, float]]:
        """
        Generate 3D point cloud from text input.
        
        Args:
            text_input: Text description of the object
            num_points: Number of points in the generated cloud
            
        Returns:
            List of (x, y, z) coordinate tuples
        """
        # For demonstration, we'll use a simple text-to-embedding approach
        # In practice, you'd use CLIP or similar
        
        # Simple text embedding (in practice, use CLIP)
        text_embedding = self._simple_text_embedding(text_input)
        
        # Generate point cloud
        with torch.no_grad():
            point_cloud = self.model(text_embedding)
            point_cloud = point_cloud.squeeze(0).cpu().numpy()
        
        # Convert to list of tuples
        points = [(float(x), float(y), float(z)) for x, y, z in point_cloud]
        
        return points
    
    def _simple_text_embedding(self, text: str) -> torch.Tensor:
        """
        Create a simple text embedding.
        
        In practice, you'd use CLIP or similar pre-trained text encoder.
        """
        # Simple hash-based embedding for demonstration
        text_hash = abs(hash(text.lower())) % (2**32 - 1)  # Ensure valid seed
        embedding = np.random.RandomState(text_hash).normal(0, 1, 512)
        return torch.tensor(embedding, dtype=torch.float32).unsqueeze(0).to(self.device)
    
    def train_on_dataset(self, dataset_path: str):
        """
        Train the model on a dataset of text-point cloud pairs.
        
        Args:
            dataset_path: Path to training dataset
        """
        # TODO: Implement training loop
        print("Training not implemented yet - would need dataset of text-point cloud pairs")
    
    def save_model(self, path: str):
        """Save the trained model."""
        torch.save(self.model.state_dict(), path)
        print(f"Model saved to {path}")
    
    def load_model(self, path: str):
        """Load a pre-trained model."""
        self.model.load_state_dict(torch.load(path, map_location=self.device))
        print(f"Model loaded from {path}")

class CLIPTextEncoder:
    """
    Wrapper for CLIP text encoding (requires CLIP installation).
    """
    
    def __init__(self):
        try:
            import clip
            self.model, self.preprocess = clip.load("ViT-B/32")
            self.available = True
        except ImportError:
            print("CLIP not available. Install with: pip install clip-by-openai")
            self.available = False
    
    def encode_text(self, text: str) -> torch.Tensor:
        """Encode text using CLIP."""
        if not self.available:
            raise RuntimeError("CLIP not available")
        
        text_tokens = clip.tokenize([text])
        with torch.no_grad():
            text_features = self.model.encode_text(text_tokens)
            text_features = F.normalize(text_features, p=2, dim=1)
        
        return text_features

# Example usage and testing
def test_ml_generation():
    """Test the ML-based point cloud generation."""
    print("Testing ML-based Point Cloud Generation...")
    
    # Initialize generator
    generator = TextToPointCloudML()
    
    # Test with different text inputs
    test_texts = [
        "a red car",
        "a fluffy cat",
        "a tall tree",
        "a modern house",
        "a flying bird"
    ]
    
    for text in test_texts:
        print(f"\nGenerating point cloud for: '{text}'")
        try:
            points = generator.generate_point_cloud(text, num_points=100)
            print(f"✓ Generated {len(points)} points")
            print(f"  First 3 points: {points[:3]}")
        except Exception as e:
            print(f"✗ Failed: {e}")

if __name__ == "__main__":
    test_ml_generation()
