requirements.txt
# Python Dependencies for Text2PointCloud

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Visualization
matplotlib>=3.5.0
plotly>=5.0.0

# Machine Learning
torch>=1.12.0
torchvision>=0.13.0
clip-by-openai>=1.0

# Additional ML dependencies
scikit-learn>=1.0.0

# Text processing (for future NLP features)
nltk>=3.7

# Development and testing
pytest>=6.0.0
pytest-cov>=3.0.0

# Code formatting and linting
black>=22.0.0
flake8>=4.0.0
